requirements.txt
streamlit>=1.50.0
pydantic>=2.11.0
deepdiff>=8.6.1
streamlit-pdf-viewer>=0.0.5
PyYAML>=6.0.3
python-dateutil>=2.8.0
pandas>=2.3.0
PyPDF2>=3.0.0
pytest>=8.4.0
pytest-mock>=3.15.0
pytest-cov>=7.0.0
coverage>=7.10.0